requirements.txt
pyspark
psycopg2-binary
echo "python-dotenv" >> requirements.txt
